feat : seperate carousel component
