+++ cypress.env.json
@@ -1,0 +1,3 @@
+{
+  "SESSION_TOKEN": "eyJhbGciOiJkaXIiLCJlbmMiOiJBMjU2R0NNIn0..jIZd89G3f6MqofW9.M4Y8PfCB34CMrc1bCKWdEqmciVL7NLqdOMX7IsdvpSsDkKyFtF7AkG9PGgEUtRxvJI0hhFhkuUer4c-rNpZvJ3Jxxi1PRY-h7_jGseTLFw3NlSZO3dFHO28EbSiN0UarsAgR0g2Tr2s_esYJaDefq42E3oYEd9IS-CNIUy9UkMqZMRlxxWDw16t0Ci35MPsSL1u4gxjwotuseFnZDo8NHsXEdeoVngD1I2JqcZ1eyZy6iuGd5o-XrsA-cq-aXlbMOVkid9OEhXUMbi2sh71b28vTmC5jdKu_XOiwNX70Jz6rt384IdApQuG7BuDNRc7wF6-7GlXLrjDbYhjQzDzW1Zz-sDQ.QQsc2bc5Iibvr_hcO75qxQ"
+}
