+++ src/hooks/index.ts
@@ -2,3 +2,4 @@ export * from './usePreview';
 export * from './useHandleArticle';
 export * from './useDetectResolution';
 export * from './useHandleOpenModal';
+export * from './useViewportWidth';
